review fixes

+++ with_numpy_types/NBond.py
@@ -7,7 +7,7 @@ class NBond(AbstractBond):
     def __init__(self, day, name, price, count):
         self._day = np.int32(day)
         self._name = np.str(name)
-        self._price = np.int32(price)
+        self._price = np.float32(price)
         self._count = np.int32(count)
 
         self._nominal = np.int32(1000)
